+++ PHY/include/pbch.h
@@ -10,5 +10,6 @@
 #include "config_parser.h"
 
 void run_pbch_simulation(const L1Config *cfg);
+void run_pbch_fading_simulation(const L1Config *cfg);
 
 #endif