+++ PHY/include/pdcch.h
@@ -10,5 +10,6 @@
 #include "config_parser.h"
 
 void run_pdcch_simulation(const L1Config *cfg);
+void run_pdcch_fading_simulation(const L1Config *cfg);
 
 #endif